making entrypoint into a function
