+++ isovar/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2016-2018. Mount Sinai School of Medicine
+# Copyright (c) 2016-2019. Mount Sinai School of Medicine
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,4 +14,8 @@
 
 from __future__ import print_function, division, absolute_import
 
-__version__ = "0.9.0"
+__version__ = "1.0.0"
+
+
+from .main import run_isovar
+from .dataframe_helpers import isovar_results_to_dataframe
